+++ analyzer/windows/modules/packages/ppt2016.py
@@ -12,7 +12,6 @@ class PPT2007(Package):
     def __init__(self, options={}, config=None):
         self.config = config
         self.options = options
-        self.options["exclude-apis"] = "memcpy"
 
     PATHS = [
         ("ProgramFiles", "Microsoft Office*", "root", "Office16", "POWERPNT.EXE"),