+++ analyzer/windows/modules/packages/pub.py
@@ -29,7 +29,6 @@ class PUB(Package):
     def __init__(self, options={}, config=None):
         self.config = config
         self.options = options
-        self.options["exclude-apis"] = "memcpy"
 
     PATHS = [
         ("ProgramFiles", "Microsoft Office", "MSPUB.EXE"),