+++ analyzer/windows/modules/packages/xls.py
@@ -14,7 +14,6 @@ class XLS(Package):
     def __init__(self, options={}, config=None):
         self.config = config
         self.options = options
-        self.options["exclude-apis"] = "memcpy"
 
     PATHS = [
         ("ProgramFiles", "Microsoft Office", "EXCEL.EXE"),